Fix sendAll on Windows.
This commit fixes some lookup problems Windows users were facing when
using the library. The culprit (as confirmed by some tests) was in
the implementation of `sendAll`, that on Windows was defined as part of
the code of this library, but this is not necessary anymore unless an
old version of GHC (< 7) is being used.

For modern environments, Network.Socket.ByteString.sendAll is all we
need.