Drop unexpected UDP answers
Occassionally, when a client gives on a query and closes its UDP
socket, the nameserver nevertheless eventually replies.  When that
happens, the UDP socket in question may now be in use for a different
query.  When the unexpected stale answer arrives, we should just
drop it, and continue waiting for the right answer.  Otherwise, we
may end up with spurious sequence number mismatch errors (observed
in practice under heavy load, with thousands of DNS queries per
second).

We check both the sequence number and the question.  For this the
question domain needs to be in standard form, with a trailing '.'.
We might have checked the question class, but questions are at
present implicitly in class "IN".  Sprinkled in some comments in
case that ever changes.

Turned the question into a singleton list early, simplifying
downsream code.