A bit of strictness and some code cleanup
A rule of thumb is that record fields should be strict.  Applied
this to the header of the DNSMessage, which will always be evaluated.
Added a few strictness annotations in the decoder, and a helper
function for parsing variable-length objects that don't have an
explicit length, but just extend to the end of the RData record.