Add rrlength to ResourceRecord.
This field may help identify inefficient or malicious encodings.

When decoding, rrlength tracks the length of the payload of
the resource record as specified in the resource record header,
regardless of whether that length is the minimum necessary.

Encoding ignores rrlength; when you construct a ResourceRecord
for the purpose of encoding it, you may set rrlength to any value.