Cabal 2.0 internal-lib reorg
Switch to a "Simple" build and create a dns-internal library.
This makes it possible to avoid exposing internal modules.

It also makes it possible (with enough effort) to run Windows
doctests, but they hang intermittently under AppVeyor, so they are
disabled on Windows, and also for GHC < 8.4.0, where the Cabal
package integration with GHC makes it hard for doctests to
find the required dependencies.

The doctests still run with GHC >= 8.4 on Linux.