Work around stack sandbox differences
Stack sanboxes and Cabal sanboxes behave differently for internal
packages. In the case of stack the internal package is hidden by
default, and the package-id values are very different.

So to make the "cbits" visible to doctests on Windows we needs to
link with just the main library, not the internal library which is
rather difficult to specify portably between cabal and stack.

Not having access to the internal library has some drawbacks, We
need to expose the internal library to the doctests in source form,
which result in those tests running twice when doctest looks for
modules to test, it finds the internal modules both directly and
indirectly...

Don't know of a better solution for now. It would be great if stack
and cabal were better aligned here.