Portably include package=dns in doctests
This configuration passes "cabal v1-test", "cabal v2-test" and
"stack test", and *builds* and *runs* doctests on Windows, which
then fail, but at least the "cbits" code is now loaded, so it could
start working later, when the source of the hang is resolved.

Sadly, it precludes using  "internal" libraries, which are not
compatible with a Custom build at present.

Which do we want more?  An internal library or "cabal v1-test"?  By
using "cabal-doctest" all the required package-db, ... flags are
passed in, making "-package=dns" work.