Retry w/o EDNS on empty FormatErr responses.
Some non-EDNS servers in response to EDNS(0) queries reply with a
zero question count.  We don't want to just drop these responses
and retry with the identical request, since then we'll never switch
to a non-EDNS query.

We therefore now accept FormatErr with an empty question section,
so long as the transaction id matches.